requirements.txt
torch
numpy
pandas
scikit-learn
pyyaml
einops
tqdm
